+++ client/src/app/remote/error-utils.ts
@@ -1,0 +1,15 @@
+import {Response} from '@angular/http';
+import {Observable} from 'rxjs';
+
+export function handleError(error: Response | any) {
+    let errMsg: string;
+    if (error instanceof Response) {
+        const body = error.json() || '';
+        const err = body.error || JSON.stringify(body);
+        errMsg = `${error.status} - ${error.statusText || ''}: ${err}`;
+    } else {
+        errMsg = error.message ? error.message : error.toString();
+    }
+    console.error(errMsg);
+    return Observable.throw(errMsg);
+}
