+++ server/routes/user.route.ts
@@ -10,7 +10,7 @@ userRoute.post('/api/users', function(req: express.Request, res: express.Respons
     console.log(`create user: ${jsonBody}`);
     user.save((err: any, addedUser: IUserModel) => {
        if (err) {
-           res.status(500).json({error: `Error creating user ${user.username}. ${err}`});
+           res.status(500).json({error: `error creating user ${user.username}. ${err}`});
        } else {
            // set the user.id to the _id provided by the db
            addedUser.id = addedUser._id;
@@ -25,7 +25,7 @@ userRoute.put('/api/users/:id', function(req: express.Request, res: express.Resp
     console.log(`update user [${id}]: ${jsonBody}`);
     User.findById(id,  (err: any, user: IUserModel) => {
         if (err) {
-            res.status(404).json({error: `User ${id} not found. ${err}`});
+            res.status(404).json({error: `user ${id} not found. ${err}`});
         } else {
             // copy the properties
             user.id = req.body.id;
@@ -37,7 +37,7 @@ userRoute.put('/api/users/:id', function(req: express.Request, res: express.Resp
             // save the updated user
             user.save((err: any, updatedUser: IUserModel) => {
                 if (err) {
-                    res.status(500).json({error: `Error updating user ${id}. ${err}`});
+                    res.status(500).json({error: `error updating user ${id}. ${err}`});
                 } else {
                     res.json(updatedUser);
                 }
@@ -49,7 +49,7 @@ userRoute.put('/api/users/:id', function(req: express.Request, res: express.Resp
 userRoute.get('/api/users', function(req: express.Request, res: express.Response, next: express.NextFunction){
     User.find((err, users) => {
        if (err) {
-           res.status(404).json({error: `Error retrieving users. ${err}`});
+           res.status(404).json({error: `error retrieving users. ${err}`});
        } else {
            // set the user.id to the _id provided by the db
            users.forEach((user) => user.id = user._id);
@@ -62,7 +62,7 @@ userRoute.get('/api/users/:id', function(req: express.Request, res: express.Resp
     let ref = {_id: req.params.id};
     User.findById(ref, (err, user) => {
         if (err) {
-            res.status(404).json({error: `Error retrieving user ${ref._id}. ${err}`});
+            res.status(404).json({error: `error retrieving user ${ref._id}. ${err}`});
         } else {
             // set the user.id to the _id provided by the db
             user.id = user._id;
@@ -75,7 +75,7 @@ userRoute.delete('/api/users/:id', function(req: express.Request, res: express.R
     let ref = {_id: req.params.id};
     User.remove( ref, (err) => {
         if (err) {
-            res.status(404).json({error: `Error deleting user ${ref._id}. ${err}`});
+            res.status(404).json({error: `error deleting user ${ref._id}. ${err}`});
         }
         res.json(ref._id);
     });
